Refactoring: added ValuesFileParser class with a Template Method pattern case to capture the commonality between DelimiterSeparatedValuesFileParser and PositionalValuesFileParser. All the tests passed untouched.
