+++ IllyumL2T.Core/IllyumL2T.Core.csproj
@@ -50,6 +50,7 @@
     <Compile Include="Parse\DelimiterSeparatedValuesFileParser.cs" />
     <Compile Include="Parse\ParseResult.cs" />
     <Compile Include="Parse\PositionalValuesFileParser.cs" />
+    <Compile Include="Parse\ValuesFileParser.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="FieldsSplit\DelimiterSeparatedValuesFieldsSplitter.cs" />
   </ItemGroup>
